push first

+++ Membership/API/Members/Payments/IMemberPayment.vb
@@ -3,8 +3,8 @@
 
     Property memberID As Integer
     Property chargesPaid As HashSet(Of IMemberCharge)
-    Property additionalCharges As Dictionary(Of String, Double)
-    Property totalPayable As Double
-    Property amountPaid As Double
-    Property balance As Double
+    Property additionalCharges As Dictionary(Of String, Decimal)
+    Property totalPayable As Decimal
+    Property amountPaid As Decimal
+    Property balance As Decimal
 End Interface
